Added the Zone class and Record class to ultradns.py and removed the respective files
